+++ wren-ai-service/src/pipelines/generation/sql_correction.py
@@ -13,7 +13,7 @@ from pydantic import BaseModel
 from src.core.engine import Engine
 from src.core.pipeline import BasicPipeline
 from src.core.provider import LLMProvider
-from src.pipelines.common import (
+from src.pipelines.generation.utils.sql import (
     TEXT_TO_SQL_RULES,
     SQLGenPostProcessor,
     sql_generation_system_prompt,
