+++ wren-ai-service/src/pipelines/generation/sql_regeneration.py
@@ -12,7 +12,7 @@ from pydantic import BaseModel
 from src.core.engine import Engine
 from src.core.pipeline import BasicPipeline
 from src.core.provider import LLMProvider
-from src.pipelines.common import SQLBreakdownGenPostProcessor
+from src.pipelines.generation.utils.sql import SQLBreakdownGenPostProcessor
 from src.utils import async_timer, timer
 from src.web.v1.services.sql_regeneration import (
     SQLExplanationWithUserCorrections,
